+++ src/components/DashboardLayout/TodayAndMonthLayout/TodayAndMonthLayout.jsx
@@ -1,7 +1,7 @@
 export default function TodayAndMonthLayout() {
   return (
-    <>
+    <div>
       <h2>TodayAndMonthLayout</h2>
-    </>
+    </div>
   );
 }
