+++ src/components/DashboardLayout/WaterRatioPanel/WaterRatioPanel.jsx
@@ -1,7 +1,7 @@
 export default function WaterRatioPanel() {
   return (
-    <>
+    <div>
       <h2>WaterRatioPanel</h2>
-    </>
+    </div>
   );
 }
