+++ src/pages/Dashboard.jsx
@@ -1,0 +1,5 @@
+import { DashboardLayout } from 'components';
+
+export default function Dashboard() {
+  return <DashboardLayout />;
+}
